Deal with playlists in search results
This commit includes a workaround for an apparent bug in the YouTube
API. Even though we specify `type=video` in the search options, we
sometimes get a playlist back as a search result. So, instead of
assuming all results are videos, check the type of the results and look
at up to 3 of them to try to find the first video.